added after_install

+++ scripts/after_install.sh
@@ -1,1 +1,3 @@
-sudo rm -R /home/ubuntu/project/scripts
+if [ -d /home/ubuntu/project/scripts ]; then
+  sudo rm -R /home/ubuntu/project/scripts
+fi
